voice-order-fastapi/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
python-dotenv==1.0.1
openai==1.51.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.4.0
python-multipart==0.0.9
